requirements.txt
numpy
scikit-learn
PyYAML
moviepy
opencv-python
librosa
openai-whisper
ffmpeg-python
sentence-transformers
openai
python-dotenv
pytest
pytest-cov
